+++ options/pin.js
@@ -1,10 +1,13 @@
+import chalk from 'chalk'
 import { getPackageJson, savePackageJson } from '../helpers/files.js'
 
-export const pinPackageManager = async (packageCommand) => {
+export const pinPackageManager = async ({ cmd, config }) => {
   const packageJson = await getPackageJson()
+
   if (packageJson) {
-    packageJson.swpm = packageCommand
+    packageJson.swpm = cmd
     await savePackageJson(packageJson)
+    console.log(`${chalk.green.bold('Success')}: ${chalk.hex(config.color).bold(cmd)} was pinned on ${chalk.bold('package.json')} file.`)
+    process.exit(0)
   }
-  process.exit(1)
 }