+++ options/test.js
@@ -1,0 +1,6 @@
+import { showCommand } from '../helpers/cmd.js'
+
+export const testCommand = async (pkg) => {
+  await showCommand(pkg)
+  process.exit(0)
+}